+++ plt/plt-scheduler/src/lib.rs
@@ -1,4 +1,4 @@
-mod block_state_polymorph;
+pub mod block_state_polymorph;
 #[cfg(feature = "ffi")]
 mod ffi;
 mod locks;
