+++ plt/plt-scheduler/src/scheduler.rs
@@ -2,12 +2,10 @@
 //! transaction and update instruction payloads.
 
 use plt_block_state::failure::BlockStateFailure;
-use plt_scheduler_types::types::reject_reasons::TransactionRejectReason;
 
 pub mod p11;
 pub mod p9;
-// todo ar
-// mod plt_scheduler;
+mod plt_scheduler;
 
 /// Unrecoverable error executing transaction. This represents the
 /// return value of [`execute_transaction`] for transactions that cannot
@@ -20,27 +18,6 @@ pub enum TransactionExecutionError {
     #[error("Block state failure: {0}")]
     BlockStateFailure(#[from] BlockStateFailure),
 }
-
-// todo ar what to do with TransactionFailure
-/// Failure mode for executing a transaction. A transaction may either be
-/// reject, which means it is included in the block but does not have any effect,
-/// or it may fail with an unrecoverable error, leading to a panic.
-#[derive(Debug, thiserror::Error)]
-pub enum TransactionFailure {
-    /// The transaction was rejected, but can be included in a block.
-    #[error("transaction rejected")]
-    Reject(TransactionRejectReason),
-    /// An unrecoverable error occurred when executing the transaction.
-    #[error("unrecoverable: {0}")]
-    Error(#[from] TransactionExecutionError),
-}
-
-impl From<TransactionRejectReason> for TransactionFailure {
-    fn from(value: TransactionRejectReason) -> Self {
-        Self::Reject(value)
-    }
-}
-
 
 /// Account with given id does not exist
 #[derive(Debug, thiserror::Error)]
